Move tests from wallet to transaction
