+++ src/wallet/service_test.go
@@ -1,29 +1,21 @@
 package wallet
 
 import (
-	"bytes"
 	"encoding/hex"
-	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"math"
 	"math/rand"
 	"os"
 	"path/filepath"
-	"reflect"
-	"sort"
 	"strings"
 	"testing"
-	"time"
 
-	"github.com/shopspring/decimal"
 	"github.com/stretchr/testify/require"
 
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/skycoin/src/coin"
 	"github.com/skycoin/skycoin/src/testutil"
-	"github.com/skycoin/skycoin/src/util/fee"
 )
 
 func prepareWltDir() string {
@@ -610,1037 +602,6 @@ func TestServiceGetWallets(t *testing.T) {
 					require.Equal(t, w, ww)
 				}
 			})
-		}
-	}
-}
-
-func TestServiceCreateTransaction(t *testing.T) {
-	headTime := uint64(time.Now().UTC().Unix())
-	seed := []byte("seed")
-
-	// Generate first keys
-	_, secKeys := cipher.MustGenerateDeterministicKeyPairsSeed(seed, 11)
-	secKey := secKeys[0]
-	addr := cipher.MustAddressFromSecKey(secKey)
-
-	var extraWalletAddrs []cipher.Address
-	for _, s := range secKeys[1:] {
-		extraWalletAddrs = append(extraWalletAddrs, cipher.MustAddressFromSecKey(s))
-	}
-
-	// Create unspent outputs
-	var uxouts []coin.UxOut
-	var originalUxouts []coin.UxOut
-	addrs := []cipher.Address{}
-	for i := 0; i < 10; i++ {
-		uxout := makeUxOut(t, secKey, 2e6, uint64(100+i))
-		uxout.Head.Time = headTime
-		uxouts = append(uxouts, uxout)
-		originalUxouts = append(originalUxouts, uxout)
-
-		a := testutil.MakeAddress()
-		addrs = append(addrs, a)
-	}
-
-	// shuffle the uxouts to test that the uxout sorting during spend selection is working
-	rand.Shuffle(len(uxouts), func(i, j int) {
-		uxouts[i], uxouts[j] = uxouts[j], uxouts[i]
-	})
-
-	// Create extra unspent outputs. These have the same value as uxouts, but are spendable by
-	// keys held in extraWalletAddrs
-	extraUxouts := make([][]coin.UxOut, len(extraWalletAddrs))
-	for j := range extraWalletAddrs {
-		s := secKeys[j+1]
-
-		var uxouts []coin.UxOut
-		for i := 0; i < 10; i++ {
-			uxout := makeUxOut(t, s, 2e6, uint64(100+i))
-			uxout.Head.Time = headTime
-			uxouts = append(uxouts, uxout)
-		}
-
-		extraUxouts[j] = uxouts
-	}
-
-	// Create unspent outputs with no hours
-	var uxoutsNoHours []coin.UxOut
-	for i := 0; i < 10; i++ {
-		uxout := makeUxOut(t, secKey, 2e6, 0)
-		uxout.Head.Time = headTime
-		uxoutsNoHours = append(uxoutsNoHours, uxout)
-	}
-
-	// shuffle the uxouts to test that the uxout sorting during spend selection is working
-	rand.Shuffle(len(uxoutsNoHours), func(i, j int) {
-		uxoutsNoHours[i], uxoutsNoHours[j] = uxoutsNoHours[j], uxoutsNoHours[i]
-	})
-
-	changeAddress := testutil.MakeAddress()
-
-	validParams := CreateTransactionParams{
-		HoursSelection: HoursSelection{
-			Type: HoursSelectionTypeManual,
-		},
-		ChangeAddress: &changeAddress,
-		To: []coin.TransactionOutput{
-			{
-				Address: addrs[0],
-				Hours:   10,
-				Coins:   1e6,
-			},
-		},
-	}
-
-	validParamsWithPassword := validParams
-	validParamsWithPassword.Wallet.Password = []byte("password")
-
-	newShareFactor := func(a string) *decimal.Decimal {
-		d, err := decimal.NewFromString(a)
-		require.NoError(t, err)
-		return &d
-	}
-
-	firstAddress := func(uxa coin.UxArray) cipher.Address {
-		require.NotEmpty(t, uxa)
-
-		addresses := make([]cipher.Address, len(uxa))
-		for i, a := range uxa {
-			addresses[i] = a.Body.Address
-		}
-
-		sort.Slice(addresses, func(i, j int) bool {
-			x := addresses[i].Bytes()
-			y := addresses[j].Bytes()
-			return bytes.Compare(x, y) < 0
-		})
-
-		return addresses[0]
-	}
-
-	cases := []struct {
-		name             string
-		err              error
-		params           CreateTransactionParams
-		opts             Options
-		unspents         []coin.UxOut
-		addressUnspents  coin.AddressUxOuts
-		chosenUnspents   []coin.UxOut
-		headTime         uint64
-		disableWalletAPI bool
-		walletNotExist   bool
-		changeOutput     *coin.TransactionOutput
-		toExpectedHours  []uint64
-	}{
-		{
-			name:             "wallet api disabled",
-			disableWalletAPI: true,
-			err:              ErrWalletAPIDisabled,
-		},
-
-		{
-			name:   "params invalid",
-			params: CreateTransactionParams{},
-			err:    NewError(errors.New("To is required")),
-		},
-
-		{
-			name:           "wallet doesn't exist",
-			params:         validParams,
-			walletNotExist: true,
-			err:            ErrWalletNotExist,
-		},
-
-		{
-			name:   "wallet encrypted and password not provided",
-			params: validParams,
-			opts: Options{
-				Encrypt: true,
-			},
-			err: ErrMissingPassword,
-		},
-
-		{
-			name:   "wallet not encrypted and password provided",
-			params: validParamsWithPassword,
-			opts: Options{
-				Encrypt: false,
-			},
-			err: ErrWalletNotEncrypted,
-		},
-
-		{
-			name: "overflowing coin hours in params",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   math.MaxUint64,
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[1],
-						Hours:   1,
-						Coins:   1e6,
-					},
-				},
-			},
-			err: NewError(errors.New("total output hours error: uint64 addition overflow")),
-		},
-
-		{
-			name: "overflowing coins in params",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   10,
-						Coins:   math.MaxUint64,
-					},
-					{
-						Address: addrs[1],
-						Hours:   1,
-						Coins:   1,
-					},
-				},
-			},
-			err: NewError(errors.New("total output coins error: uint64 addition overflow")),
-		},
-
-		{
-			name: "no unspents",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   10,
-						Coins:   1e6,
-					},
-				},
-			},
-			err: ErrNoUnspents,
-		},
-
-		{
-			name: "insufficient coins",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   10,
-						Coins:   100e6,
-					},
-				},
-			},
-			unspents: uxouts[:1],
-			err:      ErrInsufficientBalance,
-		},
-
-		{
-			name: "insufficient hours",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   100e6,
-						Coins:   1e6,
-					},
-				},
-			},
-			unspents: uxouts[:1],
-			err:      ErrInsufficientHours,
-		},
-
-		{
-			name: "insufficient coins for specified uxouts",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				Wallet: CreateTransactionWalletParams{
-					UxOuts: []cipher.SHA256{
-						extraUxouts[0][0].Hash(),
-					},
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   1,
-						Coins:   3e6,
-					},
-				},
-			},
-			addressUnspents: coin.AddressUxOuts{
-				extraWalletAddrs[0]: []coin.UxOut{extraUxouts[0][0]},
-			},
-			err: ErrInsufficientBalance,
-		},
-
-		{
-			name: "insufficient hours for specified uxouts",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				Wallet: CreateTransactionWalletParams{
-					UxOuts: []cipher.SHA256{
-						extraUxouts[0][0].Hash(),
-					},
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   200,
-						Coins:   1e6,
-					},
-				},
-			},
-			addressUnspents: coin.AddressUxOuts{
-				extraWalletAddrs[0]: []coin.UxOut{extraUxouts[0][0]},
-			},
-			err: ErrInsufficientHours,
-		},
-
-		{
-			name: "manual, 1 output, no change",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   2e6,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0]},
-		},
-
-		{
-			name: "manual, 1 output, no change, unknown address in auxs",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				Wallet: CreateTransactionWalletParams{},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   2e6,
-					},
-				},
-			},
-			addressUnspents: coin.AddressUxOuts{
-				testutil.MakeAddress(): []coin.UxOut{extraUxouts[0][0]},
-			},
-			err: ErrUnknownAddress,
-		},
-
-		{
-			name: "manual, 1 output, change",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   2e6 + 1,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1]},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   50,
-				Coins:   2e6 - 1,
-			},
-		},
-
-		{
-			name: "manual, 1 output, change, unspecified change address",
-			params: CreateTransactionParams{
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   2e6 + 1,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1]},
-			changeOutput: &coin.TransactionOutput{
-				Address: firstAddress([]coin.UxOut{originalUxouts[0], originalUxouts[1]}),
-				Hours:   50,
-				Coins:   2e6 - 1,
-			},
-		},
-
-		{
-			// there are leftover coin hours and an additional input is added
-			// to force change to save the leftover coin hours
-			name: "manual, 1 output, forced change",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   0,
-						Coins:   2e6 * 2,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1], originalUxouts[2]},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   151,
-				Coins:   2e6,
-			},
-		},
-
-		{
-			// there are leftover coin hours and no coins change,
-			// but there are no more unspents to use to force a change output
-			name: "manual, 1 output, forced change rejected no more unspents",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   80,
-						Coins:   2e6 * 2,
-					},
-				},
-			},
-			unspents:       originalUxouts[:2],
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1]},
-			changeOutput:   nil,
-		},
-
-		{
-			// there are leftover coin hours and no coins change,
-			// but the hours cost of saving them with an additional input is less than is leftover
-			name: "manual, 1 output, forced change rejected",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   80,
-						Coins:   2e6 * 2,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1]},
-			changeOutput:   nil,
-		},
-
-		{
-			name: "manual, multiple outputs",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   1e6 + 1,
-					},
-					{
-						Address: addrs[1],
-						Hours:   70,
-						Coins:   2e6,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1], originalUxouts[2], originalUxouts[3]},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   33,
-				Coins:   4e6 - 1,
-			},
-		},
-
-		{
-			name: "manual, multiple outputs, varied addressUnspents",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				Wallet: CreateTransactionWalletParams{},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   1e6 + 1,
-					},
-					{
-						Address: addrs[1],
-						Hours:   70,
-						Coins:   2e6,
-					},
-				},
-			},
-			addressUnspents: coin.AddressUxOuts{
-				extraWalletAddrs[0]: []coin.UxOut{extraUxouts[0][0]},
-				extraWalletAddrs[3]: []coin.UxOut{extraUxouts[3][1], extraUxouts[3][2]},
-				extraWalletAddrs[5]: []coin.UxOut{extraUxouts[5][6]},
-			},
-			chosenUnspents: []coin.UxOut{extraUxouts[0][0], extraUxouts[3][1], extraUxouts[3][2], extraUxouts[5][6]},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   34,
-				Coins:   4e6 - 1,
-			},
-		},
-
-		{
-			name: "manual, multiple uxouts, varied addressUnspents, wallet outputs specified",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				Wallet: CreateTransactionWalletParams{
-					UxOuts: []cipher.SHA256{
-						extraUxouts[0][0].Hash(),
-						extraUxouts[3][1].Hash(),
-						extraUxouts[3][2].Hash(),
-						extraUxouts[5][6].Hash(),
-
-						// this extra output is not necessary to satisfy the spend,
-						// it is included to test that when UxOuts are specified,
-						// only a subset is used
-						extraUxouts[0][8].Hash(),
-					},
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[0],
-						Hours:   50,
-						Coins:   1e6 + 1,
-					},
-					{
-						Address: addrs[1],
-						Hours:   70,
-						Coins:   2e6,
-					},
-				},
-			},
-			addressUnspents: coin.AddressUxOuts{
-				extraWalletAddrs[0]: []coin.UxOut{extraUxouts[0][0], extraUxouts[0][8]},
-				extraWalletAddrs[3]: []coin.UxOut{extraUxouts[3][1], extraUxouts[3][2]},
-				extraWalletAddrs[5]: []coin.UxOut{extraUxouts[5][6]},
-			},
-			chosenUnspents: []coin.UxOut{
-				extraUxouts[0][0],
-				extraUxouts[3][1],
-				extraUxouts[3][2],
-				extraUxouts[5][6],
-			},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   34,
-				Coins:   4e6 - 1,
-			},
-		},
-
-		{
-			name: "auto, multiple outputs, share factor 0.5",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: newShareFactor("0.5"),
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[0],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[1],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[4],
-						Coins:   1e3,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1], originalUxouts[2]},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   76,
-				Coins:   2e6 - (1e6 + 1e3),
-			},
-			toExpectedHours: []uint64{15, 30, 29, 1},
-		},
-
-		{
-			name: "auto, multiple outputs, share factor 0.5, switch to 1.0 because no change could be made",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: newShareFactor("0.5"),
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[0],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[1],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[4],
-						Coins:   1e6 - 1e3,
-					},
-					{
-						Address: addrs[4],
-						Coins:   1e3,
-					},
-				},
-			},
-			unspents:        []coin.UxOut{originalUxouts[0], originalUxouts[1], originalUxouts[2]},
-			chosenUnspents:  []coin.UxOut{originalUxouts[0], originalUxouts[1], originalUxouts[2]},
-			toExpectedHours: []uint64{25, 50, 50, 25, 1},
-		},
-
-		{
-			name: "encrypted, auto, multiple outputs, share factor 0.5",
-			opts: Options{
-				Encrypt:  true,
-				Password: []byte("password"),
-			},
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: newShareFactor("0.5"),
-				},
-				Wallet: CreateTransactionWalletParams{
-					Password: []byte("password"),
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[0],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[1],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[4],
-						Coins:   1e3,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1], originalUxouts[2]},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   76,
-				Coins:   2e6 - (1e6 + 1e3),
-			},
-			toExpectedHours: []uint64{15, 30, 29, 1},
-		},
-
-		{
-			name: "auto, multiple outputs, share factor 0",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: newShareFactor("0"),
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[0],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[1],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[4],
-						Coins:   1e3,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1], originalUxouts[2]},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   151,
-				Coins:   2e6 - (1e6 + 1e3),
-			},
-			toExpectedHours: []uint64{0, 0, 0, 0},
-		},
-
-		{
-			name: "auto, multiple outputs, share factor 1",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: newShareFactor("1"),
-				},
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Coins:   1e6,
-					},
-					{
-						Address: addrs[0],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[1],
-						Coins:   2e6,
-					},
-					{
-						Address: addrs[4],
-						Coins:   1e3,
-					},
-				},
-			},
-			unspents:       uxouts,
-			chosenUnspents: []coin.UxOut{originalUxouts[0], originalUxouts[1], originalUxouts[2]},
-			changeOutput: &coin.TransactionOutput{
-				Address: changeAddress,
-				Hours:   0,
-				Coins:   2e6 - (1e6 + 1e3),
-			},
-			toExpectedHours: []uint64{30, 60, 60, 1},
-		},
-
-		{
-			name:     "no coin hours in inputs",
-			unspents: uxoutsNoHours[:],
-			params: CreateTransactionParams{
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-				ChangeAddress: &changeAddress,
-				To: []coin.TransactionOutput{
-					{
-						Address: addrs[0],
-						Hours:   10,
-						Coins:   1e6,
-					},
-				},
-			},
-			err: fee.ErrTxnNoFee,
-		},
-	}
-
-	var cryptoTypes []CryptoType
-	for ct := range cryptoTable {
-		cryptoTypes = append(cryptoTypes, ct)
-	}
-
-	type TestResult struct {
-		Transaction *coin.Transaction
-		Inputs      []UxBalance
-		Unsigned    bool
-	}
-
-	results := make(map[string]TestResult, len(cases))
-
-	compareResult := func(t *testing.T, a, b TestResult) {
-		require.True(t, reflect.DeepEqual(a.Inputs, b.Inputs))
-		require.True(t, reflect.DeepEqual(a.Transaction.Out, b.Transaction.Out))
-		require.Equal(t, a.Transaction.InnerHash, b.Transaction.InnerHash)
-		require.Equal(t, a.Transaction.Type, b.Transaction.Type)
-		require.Equal(t, a.Transaction.Length, b.Transaction.Length)
-		require.Equal(t, len(a.Transaction.Sigs), len(b.Transaction.Sigs))
-
-		if a.Unsigned == b.Unsigned {
-			require.Equal(t, a.Transaction.Hash(), b.Transaction.Hash())
-			// Sigs have a nonce so will vary each run, unset them before comparing the whole transaction
-			require.Equal(t, len(a.Transaction.Sigs), len(b.Transaction.Sigs))
-			at := *a.Transaction
-			bt := *b.Transaction
-			at.Sigs = nil
-			bt.Sigs = nil
-			require.True(t, reflect.DeepEqual(at, bt))
-		} else {
-			require.NotEqual(t, a.Transaction.Hash(), b.Transaction.Hash())
-		}
-	}
-
-	bools := []bool{true, false}
-	for _, unsigned := range bools {
-		for _, tc := range cases {
-			cts := cryptoTypes
-			if !tc.opts.Encrypt {
-				cts = cts[:1]
-			}
-
-			for _, ct := range cts {
-				name := fmt.Sprintf("unsigned=%v crypto=%v %s", unsigned, ct, tc.name)
-				fmt.Println(name)
-				t.Run(name, func(t *testing.T) {
-					if tc.headTime == 0 {
-						tc.headTime = headTime
-					}
-
-					addrUxOuts := coin.AddressUxOuts{
-						addr: tc.unspents,
-					}
-
-					if tc.addressUnspents != nil {
-						addrUxOuts = tc.addressUnspents
-					}
-
-					unspents := make(map[cipher.SHA256]coin.UxOut)
-					for _, uxs := range addrUxOuts {
-						for _, ux := range uxs {
-							unspents[ux.Hash()] = ux
-						}
-					}
-
-					if tc.opts.Seed == "" {
-						tc.opts.Seed = string(seed)
-					}
-
-					dir := prepareWltDir()
-					s, err := NewService(Config{
-						WalletDir:       dir,
-						CryptoType:      ct,
-						EnableWalletAPI: true,
-					})
-					require.NoError(t, err)
-
-					if tc.walletNotExist {
-						tc.params.Wallet.ID = "foo.wlt"
-					} else {
-						wltName := NewWalletFilename()
-						opts := tc.opts
-						if opts.Encrypt && len(opts.Password) == 0 {
-							opts.Password = []byte("password")
-						}
-						w, err := s.CreateWallet(wltName, opts, nil)
-						require.NoError(t, err)
-
-						if !w.IsEncrypted() {
-							_, err := s.NewAddresses(w.Filename(), nil, 10)
-							require.NoError(t, err)
-
-							w, err = s.GetWallet(wltName)
-							require.NoError(t, err)
-
-							require.Equal(t, 11, len(w.Entries))
-							require.Equal(t, w.Entries[0].Address, addr)
-							for i, e := range w.Entries[1:] {
-								require.Equal(t, e.Address, extraWalletAddrs[i])
-							}
-						}
-
-						tc.params.Wallet.ID = wltName
-					}
-
-					s.enableWalletAPI = !tc.disableWalletAPI
-
-					tc.params.Unsigned = unsigned
-					txn, inputs, err := s.CreateTransaction(tc.params, addrUxOuts, tc.headTime)
-					if tc.err != nil {
-						require.Equal(t, tc.err, err)
-						return
-					}
-
-					require.NoError(t, err)
-
-					if tc.params.Unsigned {
-						err := txn.VerifyUnsigned()
-						require.NoError(t, err)
-					} else {
-						err := txn.Verify()
-						require.NoError(t, err)
-					}
-
-					require.Equal(t, len(inputs), len(txn.In))
-
-					// Checks duplicate inputs in array
-					inputsMap := make(map[cipher.SHA256]struct{})
-					for _, i := range inputs {
-						_, ok := inputsMap[i.Hash]
-						require.False(t, ok)
-						inputsMap[i.Hash] = struct{}{}
-					}
-
-					for i, inUxid := range txn.In {
-						_, ok := unspents[inUxid]
-						require.True(t, ok)
-
-						require.Equal(t, inUxid, inputs[i].Hash)
-					}
-
-					// Compare the transaction inputs
-					chosenUnspents := make([]coin.UxOut, len(tc.chosenUnspents))
-					chosenUnspentHashes := make([]cipher.SHA256, len(tc.chosenUnspents))
-					for i, u := range tc.chosenUnspents {
-						chosenUnspents[i] = u
-						chosenUnspentHashes[i] = u.Hash()
-					}
-					sort.Slice(chosenUnspentHashes, func(i, j int) bool {
-						return bytes.Compare(chosenUnspentHashes[i][:], chosenUnspentHashes[j][:]) < 0
-					})
-					sort.Slice(chosenUnspents, func(i, j int) bool {
-						h1 := chosenUnspents[i].Hash()
-						h2 := chosenUnspents[j].Hash()
-						return bytes.Compare(h1[:], h2[:]) < 0
-					})
-
-					sortedTxnIn := make([]cipher.SHA256, len(txn.In))
-					copy(sortedTxnIn[:], txn.In[:])
-
-					sort.Slice(sortedTxnIn, func(i, j int) bool {
-						return bytes.Compare(sortedTxnIn[i][:], sortedTxnIn[j][:]) < 0
-					})
-
-					require.Equal(t, chosenUnspentHashes, sortedTxnIn)
-
-					sort.Slice(inputs, func(i, j int) bool {
-						h1 := inputs[i].Hash
-						h2 := inputs[j].Hash
-						return bytes.Compare(h1[:], h2[:]) < 0
-					})
-
-					chosenUnspentsUxBalances := make([]UxBalance, len(chosenUnspents))
-					for i, o := range chosenUnspents {
-						b, err := NewUxBalance(tc.headTime, o)
-						require.NoError(t, err)
-						chosenUnspentsUxBalances[i] = b
-					}
-
-					require.Equal(t, chosenUnspentsUxBalances, inputs)
-
-					// Assign expected hours for comparison
-					var to []coin.TransactionOutput
-					to = append(to, tc.params.To...)
-
-					if len(tc.toExpectedHours) != 0 {
-						require.Equal(t, len(tc.toExpectedHours), len(to))
-						for i, h := range tc.toExpectedHours {
-							to[i].Hours = h
-						}
-					}
-
-					// Add the change output if specified
-					if tc.changeOutput != nil {
-						to = append(to, *tc.changeOutput)
-					}
-
-					// Compare transaction outputs
-					require.Equal(t, to, txn.Out)
-
-					// compare to previous result for the same test case
-					// but with different parameters (cryptotype, unsigned)
-					result := TestResult{
-						Transaction: txn,
-						Inputs:      inputs,
-						Unsigned:    tc.params.Unsigned,
-					}
-					if prevResult, ok := results[tc.name]; ok {
-						compareResult(t, prevResult, result)
-					} else {
-						results[tc.name] = result
-					}
-				})
-			}
 		}
 	}
 }
