+++ src/wallet/transaction.go
@@ -9,6 +9,13 @@ import (
 	"github.com/skycoin/skycoin/src/transaction"
 )
 
+var (
+	// ErrUnknownAddress is returned if an address is not found in a wallet
+	ErrUnknownAddress = NewError(errors.New("Address not found in wallet"))
+	// ErrUnknownUxOut is returned if a uxout is not owned by any address in a wallet
+	ErrUnknownUxOut = NewError(errors.New("UxOut is not owned by any address in the wallet"))
+)
+
 func validateSignIndexes(x []int, uxOuts []coin.UxOut) error {
 	if len(x) > len(uxOuts) {
 		return errors.New("Number of signature indexes exceeds number of inputs")
