+++ src/wallet/wallet_test.go
@@ -1,29 +1,20 @@
 package wallet
 
 import (
-	"bytes"
 	"encoding/hex"
 	"errors"
 	"flag"
 	"fmt"
 	"io/ioutil"
-	"math"
 	"math/rand"
-	"reflect"
 	"testing"
 	"time"
 
-	"github.com/shopspring/decimal"
 	"github.com/stretchr/testify/require"
 
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/skycoin/src/cipher/encrypt"
-	"github.com/skycoin/skycoin/src/coin"
-	"github.com/skycoin/skycoin/src/params"
-	"github.com/skycoin/skycoin/src/testutil"
-	"github.com/skycoin/skycoin/src/util/fee"
 	"github.com/skycoin/skycoin/src/util/logging"
-	"github.com/skycoin/skycoin/src/util/mathutil"
 )
 
 var (
@@ -859,685 +850,6 @@ func TestWalletGuard(t *testing.T) {
 	}
 }
 
-type distributeSpendHoursTestCase struct {
-	name              string
-	inputHours        uint64
-	nAddrs            uint64
-	haveChange        bool
-	expectChangeHours uint64
-	expectAddrHours   []uint64
-}
-
-var burnFactor2TestCases = []distributeSpendHoursTestCase{
-	{
-		name:            "no input hours, one addr, no change",
-		inputHours:      0,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "no input hours, two addrs, no change",
-		inputHours:      0,
-		nAddrs:          2,
-		haveChange:      false,
-		expectAddrHours: []uint64{0, 0},
-	},
-	{
-		name:            "no input hours, one addr, change",
-		inputHours:      0,
-		nAddrs:          1,
-		haveChange:      true,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "one input hour, one addr, no change",
-		inputHours:      1,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "two input hours, one addr, no change",
-		inputHours:      2,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{1},
-	},
-	{
-		name:              "two input hours, one addr, change",
-		inputHours:        2,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{0},
-	},
-	{
-		name:              "three input hours, one addr, change",
-		inputHours:        3,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{0},
-	},
-	{
-		name:            "three input hours, one addr, no change",
-		inputHours:      3,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{1},
-	},
-	{
-		name:            "three input hours, two addrs, no change",
-		inputHours:      3,
-		nAddrs:          2,
-		haveChange:      false,
-		expectAddrHours: []uint64{1, 0},
-	},
-	{
-		name:            "four input hours, one addr, no change",
-		inputHours:      4,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{2},
-	},
-	{
-		name:              "four input hours, one addr, change",
-		inputHours:        4,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{1},
-	},
-	{
-		name:              "four input hours, two addr, change",
-		inputHours:        4,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{1, 0},
-	},
-	{
-		name:              "30 (divided by 2, odd number) input hours, two addr, change",
-		inputHours:        30,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 8,
-		expectAddrHours:   []uint64{4, 3},
-	},
-	{
-		name:              "33 (odd number) input hours, two addr, change",
-		inputHours:        33,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 8,
-		expectAddrHours:   []uint64{4, 4},
-	},
-	{
-		name:              "33 (odd number) input hours, three addr, change",
-		inputHours:        33,
-		nAddrs:            3,
-		haveChange:        true,
-		expectChangeHours: 8,
-		expectAddrHours:   []uint64{3, 3, 2},
-	},
-}
-
-var burnFactor3TestCases = []distributeSpendHoursTestCase{
-	{
-		name:            "no input hours, one addr, no change",
-		inputHours:      0,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "no input hours, two addrs, no change",
-		inputHours:      0,
-		nAddrs:          2,
-		haveChange:      false,
-		expectAddrHours: []uint64{0, 0},
-	},
-	{
-		name:            "no input hours, one addr, change",
-		inputHours:      0,
-		nAddrs:          1,
-		haveChange:      true,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "one input hour, one addr, no change",
-		inputHours:      1,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "two input hours, one addr, no change",
-		inputHours:      2,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{1},
-	},
-	{
-		name:            "three input hours, one addr, no change",
-		inputHours:      3,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{2},
-	},
-	{
-		name:              "two input hours, one addr, change",
-		inputHours:        2,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{0},
-	},
-	{
-		name:              "three input hours, one addr, change",
-		inputHours:        3,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{1},
-	},
-	{
-		name:              "four input hours, one addr, change",
-		inputHours:        4,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{1},
-	},
-	{
-		name:            "four input hours, one addr, no change",
-		inputHours:      4,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{2},
-	},
-	{
-		name:            "four input hours, two addrs, no change",
-		inputHours:      4,
-		nAddrs:          2,
-		haveChange:      false,
-		expectAddrHours: []uint64{1, 1},
-	},
-	{
-		name:            "five input hours, one addr, no change",
-		inputHours:      5,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{3},
-	},
-	{
-		name:              "five input hours, one addr, change",
-		inputHours:        5,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 2,
-		expectAddrHours:   []uint64{1},
-	},
-	{
-		name:              "five input hours, two addr, change",
-		inputHours:        5,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 2,
-		expectAddrHours:   []uint64{1, 0},
-	},
-	{
-		name:              "32 input hours, two addr, change",
-		inputHours:        32,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 11,
-		expectAddrHours:   []uint64{5, 5},
-	},
-	{
-		name:              "35 input hours, two addr, change",
-		inputHours:        35,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 12,
-		expectAddrHours:   []uint64{6, 5},
-	},
-	{
-		name:              "32 input hours, three addr, change",
-		inputHours:        32,
-		nAddrs:            3,
-		haveChange:        true,
-		expectChangeHours: 11,
-		expectAddrHours:   []uint64{4, 3, 3},
-	},
-}
-
-var burnFactor10TestCases = []distributeSpendHoursTestCase{
-	{
-		name:            "no input hours, one addr, no change",
-		inputHours:      0,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "no input hours, two addrs, no change",
-		inputHours:      0,
-		nAddrs:          2,
-		haveChange:      false,
-		expectAddrHours: []uint64{0, 0},
-	},
-	{
-		name:            "no input hours, one addr, change",
-		inputHours:      0,
-		nAddrs:          1,
-		haveChange:      true,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "one input hour, one addr, no change",
-		inputHours:      1,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{0},
-	},
-	{
-		name:            "two input hours, one addr, no change",
-		inputHours:      2,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{1},
-	},
-	{
-		name:            "three input hours, one addr, no change",
-		inputHours:      3,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{2},
-	},
-	{
-		name:              "two input hours, one addr, change",
-		inputHours:        2,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{0},
-	},
-	{
-		name:              "three input hours, one addr, change",
-		inputHours:        3,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 1,
-		expectAddrHours:   []uint64{1},
-	},
-	{
-		name:              "four input hours, one addr, change",
-		inputHours:        4,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 2,
-		expectAddrHours:   []uint64{1},
-	},
-	{
-		name:            "four input hours, one addr, no change",
-		inputHours:      4,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{3},
-	},
-	{
-		name:            "four input hours, two addrs, no change",
-		inputHours:      4,
-		nAddrs:          2,
-		haveChange:      false,
-		expectAddrHours: []uint64{2, 1},
-	},
-	{
-		name:            "five input hours, one addr, no change",
-		inputHours:      5,
-		nAddrs:          1,
-		haveChange:      false,
-		expectAddrHours: []uint64{4},
-	},
-	{
-		name:              "five input hours, one addr, change",
-		inputHours:        5,
-		nAddrs:            1,
-		haveChange:        true,
-		expectChangeHours: 2,
-		expectAddrHours:   []uint64{2},
-	},
-	{
-		name:              "five input hours, two addr, change",
-		inputHours:        5,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 2,
-		expectAddrHours:   []uint64{1, 1},
-	},
-	{
-		name:              "32 input hours, two addr, change",
-		inputHours:        32,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 14,
-		expectAddrHours:   []uint64{7, 7},
-	},
-	{
-		name:              "35 input hours, two addr, change",
-		inputHours:        35,
-		nAddrs:            2,
-		haveChange:        true,
-		expectChangeHours: 16,
-		expectAddrHours:   []uint64{8, 7},
-	},
-	{
-		name:              "32 input hours, three addr, change",
-		inputHours:        32,
-		nAddrs:            3,
-		haveChange:        true,
-		expectChangeHours: 14,
-		expectAddrHours:   []uint64{5, 5, 4},
-	},
-}
-
-func TestWalletDistributeSpendHours(t *testing.T) {
-	originalBurnFactor := params.UserVerifyTxn.BurnFactor
-
-	cases := []struct {
-		burnFactor uint32
-		cases      []distributeSpendHoursTestCase
-	}{
-		{2, burnFactor2TestCases},
-		{3, burnFactor3TestCases},
-		{10, burnFactor10TestCases},
-	}
-
-	tested := false
-	for _, tcc := range cases {
-		if tcc.burnFactor == params.UserVerifyTxn.BurnFactor {
-			tested = true
-		}
-
-		for _, tc := range tcc.cases {
-			t.Run(tc.name, func(t *testing.T) {
-				params.UserVerifyTxn.BurnFactor = tcc.burnFactor
-				defer func() {
-					params.UserVerifyTxn.BurnFactor = originalBurnFactor
-				}()
-
-				changeHours, addrHours, totalHours := DistributeSpendHours(tc.inputHours, tc.nAddrs, tc.haveChange)
-				require.Equal(t, tc.expectChangeHours, changeHours)
-				require.Equal(t, tc.expectAddrHours, addrHours)
-				require.Equal(t, tc.nAddrs, uint64(len(addrHours)))
-
-				outputHours := changeHours
-				for _, h := range addrHours {
-					outputHours += h
-				}
-				require.True(t, tc.inputHours >= outputHours)
-				require.Equal(t, outputHours, totalHours)
-
-				if tc.inputHours != 0 {
-					err := fee.VerifyTransactionFeeForHours(outputHours, tc.inputHours-outputHours, params.UserVerifyTxn.BurnFactor)
-					require.NoError(t, err)
-				}
-			})
-		}
-
-		t.Run(fmt.Sprintf("burn-factor-%d-range", tcc.burnFactor), func(t *testing.T) {
-			params.UserVerifyTxn.BurnFactor = tcc.burnFactor
-			defer func() {
-				params.UserVerifyTxn.BurnFactor = originalBurnFactor
-			}()
-
-			// Tests over range of values
-			for inputHours := uint64(0); inputHours <= 1e3; inputHours++ {
-				for nAddrs := uint64(1); nAddrs < 16; nAddrs++ {
-					for _, haveChange := range []bool{true, false} {
-						name := fmt.Sprintf("inputHours=%d nAddrs=%d haveChange=%v", inputHours, nAddrs, haveChange)
-						t.Run(name, func(t *testing.T) {
-							changeHours, addrHours, totalHours := DistributeSpendHours(inputHours, nAddrs, haveChange)
-							require.Equal(t, nAddrs, uint64(len(addrHours)))
-
-							var sumAddrHours uint64
-							for _, h := range addrHours {
-								sumAddrHours += h
-							}
-
-							if haveChange {
-								remainingHours := (inputHours - fee.RequiredFee(inputHours, params.UserVerifyTxn.BurnFactor))
-								splitRemainingHours := remainingHours / 2
-								require.True(t, changeHours == splitRemainingHours || changeHours == splitRemainingHours+1)
-								require.Equal(t, splitRemainingHours, sumAddrHours)
-							} else {
-								require.Equal(t, uint64(0), changeHours)
-								require.Equal(t, inputHours-fee.RequiredFee(inputHours, params.UserVerifyTxn.BurnFactor), sumAddrHours)
-							}
-
-							outputHours := sumAddrHours + changeHours
-							require.True(t, inputHours >= outputHours)
-							require.Equal(t, outputHours, totalHours)
-
-							if inputHours != 0 {
-								err := fee.VerifyTransactionFeeForHours(outputHours, inputHours-outputHours, params.UserVerifyTxn.BurnFactor)
-								require.NoError(t, err)
-							}
-
-							// addrHours at the beginning and end of the array should not differ by more than one
-							max := addrHours[0]
-							min := addrHours[len(addrHours)-1]
-							require.True(t, max-min <= 1)
-						})
-					}
-				}
-			}
-		})
-	}
-
-	require.True(t, tested, "configured BurnFactor=%d has not been tested", params.UserVerifyTxn.BurnFactor)
-}
-
-func uxBalancesEqual(a, b []UxBalance) bool {
-	if len(a) != len(b) {
-		return false
-	}
-
-	for i, x := range a {
-		if x != b[i] {
-			return false
-		}
-	}
-
-	return true
-}
-
-func TestWalletSortSpendsLowToHigh(t *testing.T) {
-	// UxBalances are sorted with Coins lowest, then following other order rules
-	orderedUxb := []UxBalance{
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 5,
-			Coins: 1,
-			Hours: 0,
-		},
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 3,
-			Coins: 10,
-			Hours: 1,
-		},
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 1,
-			Coins: 10,
-			Hours: 10,
-		},
-		{
-			Hash:  cipher.MustSHA256FromHex("bddf0aaf80f96c144f33ac8a27764a868d37e1c11e568063ebeb1367de859566"),
-			BkSeq: 2,
-			Coins: 10,
-			Hours: 10,
-		},
-		{
-			Hash:  cipher.MustSHA256FromHex("f569461182b0efe9a5c666e9a35c6602b351021c1803cc740aca548cf6db4cb2"),
-			BkSeq: 2,
-			Coins: 10,
-			Hours: 10,
-		},
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 4,
-			Coins: 100,
-			Hours: 100,
-		},
-	}
-
-	shuffleWorked := false
-	nShuffle := 20
-	for i := 0; i < nShuffle; i++ {
-		// Shuffle the list
-		uxb := make([]UxBalance, len(orderedUxb))
-		copy(uxb, orderedUxb)
-
-		for i := range uxb {
-			j := rand.Intn(i + 1)
-			uxb[i], uxb[j] = uxb[j], uxb[i]
-		}
-
-		// Sanity check that shuffling produces a new result
-		if !uxBalancesEqual(uxb, orderedUxb) {
-			shuffleWorked = true
-		}
-
-		sortSpendsCoinsLowToHigh(uxb)
-
-		for i, ux := range uxb {
-			require.Equal(t, orderedUxb[i], ux, "index %d", i)
-		}
-
-		verifySortedCoinsLowToHigh(t, uxb)
-	}
-
-	require.True(t, shuffleWorked)
-
-	nRand := 1000
-	for i := 0; i < nRand; i++ {
-		uxb := makeRandomUxBalances(t)
-
-		sortSpendsCoinsHighToLow(uxb)
-		verifySortedCoinsHighToLow(t, uxb)
-	}
-}
-
-func TestWalletSortSpendsHighToLow(t *testing.T) {
-	// UxBalances are sorted with Coins highest, then following other order rules
-	orderedUxb := []UxBalance{
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 4,
-			Coins: 10000,
-			Hours: 0,
-		},
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 10,
-			Coins: 1000,
-			Hours: 1,
-		},
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 4,
-			Coins: 100,
-			Hours: 100,
-		},
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 3,
-			Coins: 10,
-			Hours: 1,
-		},
-		{
-			Hash:  testutil.RandSHA256(t),
-			BkSeq: 1,
-			Coins: 10,
-			Hours: 10,
-		},
-		{
-			Hash:  cipher.MustSHA256FromHex("bddf0aaf80f96c144f33ac8a27764a868d37e1c11e568063ebeb1367de859566"),
-			BkSeq: 2,
-			Coins: 10,
-			Hours: 10,
-		},
-		{
-			Hash:  cipher.MustSHA256FromHex("f569461182b0efe9a5c666e9a35c6602b351021c1803cc740aca548cf6db4cb2"),
-			BkSeq: 2,
-			Coins: 10,
-			Hours: 10,
-		},
-	}
-
-	shuffleWorked := false
-	nShuffle := 20
-	for i := 0; i < nShuffle; i++ {
-		// Shuffle the list
-		uxb := make([]UxBalance, len(orderedUxb))
-		copy(uxb, orderedUxb)
-
-		rand.Shuffle(len(uxb), func(i, j int) {
-			uxb[i], uxb[j] = uxb[j], uxb[i]
-		})
-
-		if !uxBalancesEqual(uxb, orderedUxb) {
-			shuffleWorked = true
-		}
-
-		sortSpendsCoinsHighToLow(uxb)
-
-		for i, ux := range uxb {
-			require.Equal(t, orderedUxb[i], ux, "index %d", i)
-		}
-
-		verifySortedCoinsHighToLow(t, uxb)
-	}
-
-	require.True(t, shuffleWorked)
-
-	nRand := 1000
-	for i := 0; i < nRand; i++ {
-		uxb := makeRandomUxBalances(t)
-
-		sortSpendsCoinsHighToLow(uxb)
-		verifySortedCoinsHighToLow(t, uxb)
-	}
-}
-
-func TestWalletChooseSpendsMaximizeUxOuts(t *testing.T) {
-	nRand := 10000
-	for i := 0; i < nRand; i++ {
-		coins := uint64((rand.Intn(3)+1)*10 + rand.Intn(3)) // 10,20,30 + 0,1,2
-		uxb := makeRandomUxBalances(t)
-
-		verifyChosenCoins(t, uxb, coins, ChooseSpendsMaximizeUxOuts, func(a, b UxBalance) bool {
-			return a.Coins <= b.Coins
-		})
-	}
-}
-
-func TestWalletChooseSpendsMinimizeUxOutsRandom(t *testing.T) {
-	nRand := 10000
-	for i := 0; i < nRand; i++ {
-		coins := uint64((rand.Intn(3)+1)*10 + rand.Intn(3)) // 10,20,30 + 0,1,2
-		uxb := makeRandomUxBalances(t)
-
-		verifyChosenCoins(t, uxb, coins, ChooseSpendsMinimizeUxOuts, func(a, b UxBalance) bool {
-			return a.Coins >= b.Coins
-		})
-	}
-}
-
 func TestRemoveBackupFiles(t *testing.T) {
 	type wltInfo struct {
 		wltName string
@@ -1837,755 +1149,6 @@ func TestRemoveBackupFiles(t *testing.T) {
 	}
 }
 
-func makeRandomUxBalances(t *testing.T) []UxBalance {
-	// Generate random 0-100 UxBalances
-	// Coins 1-10 (must be >0)
-	// Hours 0-10
-	// BkSeq 0-10
-	// Hash random
-	// Small ranges are used for Coins, Hours, BkSeq to increase likelihood
-	// that they collide and test deeper sorting comparisons
-
-	n := rand.Intn(101)
-	uxb := make([]UxBalance, n)
-
-	// Use a random max range for the hours' rand range to ensure enough
-	// balances have zero hours
-	hasZeroHoursRange := rand.Intn(3) + 1
-
-	for i := 0; i < n; i++ {
-		ux := UxBalance{
-			Coins: uint64(rand.Intn(10) + 1), // 1-10
-			Hours: uint64(rand.Intn(hasZeroHoursRange)),
-			BkSeq: uint64(rand.Intn(11)), // 0-10
-			Hash:  testutil.RandSHA256(t),
-		}
-
-		uxb[i] = ux
-	}
-
-	return uxb
-}
-
-func verifyChosenCoins(t *testing.T, uxb []UxBalance, coins uint64, chooseSpends func([]UxBalance, uint64, uint64) ([]UxBalance, error), cmpCoins func(i, j UxBalance) bool) {
-	var haveZero, haveNonzero int
-	for _, ux := range uxb {
-		if ux.Hours == 0 {
-			haveZero++
-		} else {
-			haveNonzero++
-		}
-	}
-
-	var totalCoins, totalHours uint64
-	for _, ux := range uxb {
-		totalCoins += ux.Coins
-		totalHours += ux.Hours
-	}
-
-	chosen, err := chooseSpends(uxb, coins, 0)
-
-	if coins == 0 {
-		testutil.RequireError(t, err, ErrZeroSpend.Error())
-		return
-	}
-
-	if len(uxb) == 0 {
-		testutil.RequireError(t, err, ErrNoUnspents.Error())
-		return
-	}
-
-	if totalHours == 0 {
-		testutil.RequireError(t, err, fee.ErrTxnNoFee.Error())
-		return
-	}
-
-	if coins > totalCoins {
-		testutil.RequireError(t, err, ErrInsufficientBalance.Error())
-		return
-	}
-
-	require.NoError(t, err)
-	require.NotEqual(t, 0, len(chosen))
-
-	// Check that there are no duplicated spends chosen
-	uxMap := make(map[UxBalance]struct{}, len(chosen))
-	for _, ux := range chosen {
-		_, ok := uxMap[ux]
-		require.False(t, ok)
-		uxMap[ux] = struct{}{}
-	}
-
-	// The first chosen spend should have non-zero coin hours
-	require.NotEqual(t, uint64(0), chosen[0].Hours)
-
-	// Outputs with zero hours should come before any outputs with non-zero hours,
-	// except for the first output
-	for i := range chosen {
-		if i <= 1 {
-			continue
-		}
-
-		a := chosen[i-1]
-		b := chosen[i]
-
-		if b.Hours == 0 {
-			require.Equal(t, uint64(0), a.Hours)
-		}
-	}
-
-	// The initial UxBalance with hours should have more or equal coins than any other UxBalance with hours
-	// If it has equal coins, it should have less hours
-	for _, ux := range chosen[1:] {
-		if ux.Hours != 0 {
-			require.True(t, chosen[0].Coins >= ux.Coins)
-
-			if chosen[0].Coins == ux.Coins {
-				require.True(t, chosen[0].Hours <= ux.Hours)
-			}
-		}
-	}
-
-	var zeroBalances, nonzeroBalances []UxBalance
-	for _, ux := range chosen[1:] {
-		if ux.Hours == 0 {
-			zeroBalances = append(zeroBalances, ux)
-		} else {
-			nonzeroBalances = append(nonzeroBalances, ux)
-		}
-	}
-
-	// Amongst the UxBalances with zero hours, they should be sorted as specified
-	verifySortedCoins(t, zeroBalances, cmpCoins)
-
-	// Amongst the UxBalances with non-zero hours, they should be sorted as specified
-	verifySortedCoins(t, nonzeroBalances, cmpCoins)
-
-	// If there are any extra UxBalances with non-zero hours, all of the zeros should have been chosen
-	if len(nonzeroBalances) > 0 {
-		require.Equal(t, haveZero, len(zeroBalances))
-	}
-
-	// Excessive UxBalances to satisfy the amount requested should not be included
-	var haveCoins uint64
-	for i, ux := range chosen {
-		haveCoins += ux.Coins
-		if haveCoins >= coins {
-			require.Equal(t, len(chosen)-1, i)
-		}
-	}
-}
-
-func verifySortedCoins(t *testing.T, uxb []UxBalance, cmpCoins func(a, b UxBalance) bool) {
-	if len(uxb) <= 1 {
-		return
-	}
-
-	for i := range uxb {
-		if i == 0 {
-			continue
-		}
-
-		a := uxb[i-1]
-		b := uxb[i]
-
-		require.True(t, cmpCoins(a, b))
-
-		if a.Coins == b.Coins {
-			require.True(t, a.Hours <= b.Hours)
-
-			if a.Hours == b.Hours {
-				require.True(t, a.BkSeq <= b.BkSeq)
-
-				if a.BkSeq == b.BkSeq {
-					cmp := bytes.Compare(a.Hash[:], b.Hash[:])
-					require.True(t, cmp < 0)
-				}
-			}
-		}
-	}
-}
-
-func verifySortedCoinsLowToHigh(t *testing.T, uxb []UxBalance) {
-	verifySortedCoins(t, uxb, func(a, b UxBalance) bool {
-		return a.Coins <= b.Coins
-	})
-}
-
-func verifySortedCoinsHighToLow(t *testing.T, uxb []UxBalance) {
-	verifySortedCoins(t, uxb, func(a, b UxBalance) bool {
-		return a.Coins >= b.Coins
-	})
-}
-
-func TestCreateWalletParamsVerify(t *testing.T) {
-	changeAddress := testutil.MakeAddress()
-
-	toManual := []coin.TransactionOutput{
-		{
-			Address: testutil.MakeAddress(),
-			Coins:   1e6,
-			Hours:   1,
-		},
-		{
-			Address: testutil.MakeAddress(),
-			Coins:   5e6,
-			Hours:   0,
-		},
-	}
-
-	toAuto := []coin.TransactionOutput{
-		{
-			Address: testutil.MakeAddress(),
-			Coins:   1e6,
-		},
-		{
-			Address: testutil.MakeAddress(),
-			Coins:   5e6,
-		},
-	}
-
-	one := decimal.New(1, 0)
-	negativeOne := decimal.New(-1, 0)
-	onePointOne := decimal.New(11, -1)
-	pointOneOne := decimal.New(11, -2)
-
-	uxoutHash := testutil.RandSHA256(t)
-
-	cases := []struct {
-		name   string
-		params CreateTransactionParams
-		err    string
-	}{
-		{
-			name: "null change address",
-			params: CreateTransactionParams{
-				ChangeAddress: &cipher.Address{},
-			},
-			err: "ChangeAddress must not be the null address",
-		},
-
-		{
-			name: "no to destinations",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-			},
-			err: "To is required",
-		},
-
-		{
-			name: "missing to coins",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To: []coin.TransactionOutput{
-					{
-						Address: testutil.MakeAddress(),
-						Hours:   1,
-					},
-				},
-			},
-			err: "To.Coins must not be zero",
-		},
-
-		{
-			name: "missing to address",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To: []coin.TransactionOutput{
-					{
-						Coins: 5,
-						Hours: 1,
-					},
-				},
-			},
-			err: "To.Address must not be the null address",
-		},
-
-		{
-			name: "missing wallet id",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-			},
-			err: "Wallet.ID is required",
-		},
-
-		{
-			name: "wallet addresses contains empty value",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{cipher.Address{}},
-				},
-			},
-			err: "Wallet.Addresses must not contain the null address",
-		},
-
-		{
-			name: "nonzero to hours for auto selection",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeAuto,
-				},
-			},
-			err: "To.Hours must be zero for auto type hours selection",
-		},
-
-		{
-			name: "mode missing for auto selection",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toAuto,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeAuto,
-				},
-			},
-			err: "HoursSelection.Mode is required for auto type hours selection",
-		},
-
-		{
-			name: "mode set for manual selection",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-					Mode: HoursSelectionModeShare,
-				},
-			},
-			err: "HoursSelection.Mode cannot be used for manual type hours selection",
-		},
-
-		{
-			name: "missing hours selection type",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toAuto,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type: "",
-				},
-			},
-			err: "Invalid HoursSelection.Type",
-		},
-
-		{
-			name: "invalid hours selection type",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toAuto,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type: "invalid",
-				},
-			},
-			err: "Invalid HoursSelection.Type",
-		},
-
-		{
-			name: "invalid hours selection mode",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toAuto,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeAuto,
-					Mode: "invalid",
-				},
-			},
-			err: "Invalid HoursSelection.Mode",
-		},
-
-		{
-			name: "share factor not set for split even mode",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toAuto,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeAuto,
-					Mode: HoursSelectionModeShare,
-				},
-			},
-			err: "HoursSelection.ShareFactor must be set for share mode",
-		},
-
-		{
-			name: "share factor set but not split even mode",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeManual,
-					ShareFactor: &one,
-				},
-			},
-			err: "HoursSelection.ShareFactor can only be used for share mode",
-		},
-
-		{
-			name: "share factor less than 0",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toAuto,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: &negativeOne,
-				},
-			},
-			err: "HoursSelection.ShareFactor must be >= 0 and <= 1",
-		},
-
-		{
-			name: "share factor greater than 1",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toAuto,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: &onePointOne,
-				},
-			},
-			err: "HoursSelection.ShareFactor must be >= 0 and <= 1",
-		},
-
-		{
-			name: "duplicate output when manual",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            []coin.TransactionOutput{toManual[0], toManual[0]},
-				Wallet: CreateTransactionWalletParams{
-					ID: "foo.wlt",
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-			},
-			err: "To contains duplicate values",
-		},
-
-		{
-			name: "duplicate output when auto",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            []coin.TransactionOutput{toAuto[0], toAuto[0]},
-				Wallet: CreateTransactionWalletParams{
-					ID: "foo.wlt",
-				},
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: &pointOneOne,
-				},
-			},
-			err: "To contains duplicate values",
-		},
-
-		{
-			name: "both uxouts and addresses specified",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{changeAddress},
-					UxOuts:    []cipher.SHA256{uxoutHash},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-			},
-			err: "Wallet.UxOuts and Wallet.Addresses cannot be combined",
-		},
-
-		{
-			name: "duplicate uxouts",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-				Wallet: CreateTransactionWalletParams{
-					ID:     "foo.wlt",
-					UxOuts: []cipher.SHA256{uxoutHash, uxoutHash},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-			},
-			err: "Wallet.UxOuts contains duplicate values",
-		},
-
-		{
-			name: "duplicate addresses",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{changeAddress, changeAddress},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-			},
-			err: "Wallet.Addresses contains duplicate values",
-		},
-
-		{
-			name: "valid auto split even share factor",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toAuto,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type:        HoursSelectionTypeAuto,
-					Mode:        HoursSelectionModeShare,
-					ShareFactor: &pointOneOne,
-				},
-			},
-		},
-
-		{
-			name: "valid manual",
-			params: CreateTransactionParams{
-				ChangeAddress: &changeAddress,
-				To:            toManual,
-				Wallet: CreateTransactionWalletParams{
-					ID:        "foo.wlt",
-					Addresses: []cipher.Address{},
-				},
-				HoursSelection: HoursSelection{
-					Type: HoursSelectionTypeManual,
-				},
-			},
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			err := tc.params.Validate()
-			if tc.err != "" {
-				require.Equal(t, NewError(errors.New(tc.err)), err, err.Error())
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestDistributeCoinHoursProportional(t *testing.T) {
-	cases := []struct {
-		name   string
-		coins  []uint64
-		hours  uint64
-		output []uint64
-		err    error
-	}{
-		{
-			name:  "no coins",
-			hours: 1,
-			err:   errors.New("DistributeCoinHoursProportional coins array must not be empty"),
-		},
-		{
-			name:  "coins have 0 in them",
-			coins: []uint64{1, 2, 0, 3},
-			hours: 1,
-			err:   errors.New("DistributeCoinHoursProportional coins array has a zero value"),
-		},
-		{
-			name:  "total coins too large while adding",
-			coins: []uint64{10, math.MaxUint64 - 9},
-			hours: 1,
-			err:   mathutil.ErrUint64AddOverflow,
-		},
-		{
-			name:  "total coins too large after adding",
-			coins: []uint64{10, math.MaxInt64},
-			hours: 1,
-			err:   mathutil.ErrUint64OverflowsInt64,
-		},
-		{
-			name:  "single coin too large",
-			coins: []uint64{10, math.MaxInt64 + 1},
-			hours: 1,
-			err:   mathutil.ErrUint64OverflowsInt64,
-		},
-		{
-			name:  "hours too large",
-			coins: []uint64{10},
-			hours: math.MaxInt64 + 1,
-			err:   mathutil.ErrUint64OverflowsInt64,
-		},
-
-		{
-			name:   "valid, one input",
-			coins:  []uint64{1},
-			hours:  1,
-			output: []uint64{1},
-		},
-
-		{
-			name:   "zero hours",
-			coins:  []uint64{1},
-			hours:  0,
-			output: []uint64{0},
-		},
-
-		{
-			name:   "valid, multiple inputs, all equal",
-			coins:  []uint64{2, 4, 8, 16},
-			hours:  30,
-			output: []uint64{2, 4, 8, 16},
-		},
-
-		{
-			name:   "valid, multiple inputs, rational division in coins and hours",
-			coins:  []uint64{2, 4, 8, 16},
-			hours:  30,
-			output: []uint64{2, 4, 8, 16},
-		},
-
-		{
-			name:   "valid, multiple inputs, rational division in coins, irrational in hours",
-			coins:  []uint64{2, 4, 8, 16},
-			hours:  31,
-			output: []uint64{3, 4, 8, 16},
-		},
-
-		{
-			name:   "valid, multiple inputs, irrational division in coins, rational in hours",
-			coins:  []uint64{2, 3, 5, 7, 11, 13},
-			hours:  41,
-			output: []uint64{2, 3, 5, 7, 11, 13},
-		},
-
-		{
-			name:   "valid, multiple inputs, irrational division in coins and hours",
-			coins:  []uint64{2, 3, 5, 7, 11, 13},
-			hours:  50,
-			output: []uint64{3, 4, 7, 8, 13, 15},
-		},
-
-		{
-			name:   "valid, multiple inputs that would receive 0 hours but get compensated from remainder as priority",
-			coins:  []uint64{16, 8, 4, 2, 1, 1},
-			hours:  14,
-			output: []uint64{7, 3, 1, 1, 1, 1},
-		},
-
-		{
-			name:   "not enough hours for everyone",
-			coins:  []uint64{1, 1, 1, 1, 1},
-			hours:  1,
-			output: []uint64{1, 0, 0, 0, 0},
-		},
-
-		{
-			name:   "not enough hours for everyone 2",
-			coins:  []uint64{1, 1, 1, 1, 1},
-			hours:  3,
-			output: []uint64{1, 1, 1, 0, 0},
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			hours, err := DistributeCoinHoursProportional(tc.coins, tc.hours)
-			if tc.err != nil {
-				require.Equal(t, tc.err, err)
-			} else {
-				require.NoError(t, err)
-				require.Equal(t, tc.output, hours)
-			}
-		})
-	}
-
-	// Randomized tests
-	iterations := 10000
-	maxCoinsLen := 300
-	maxMaxCoins := 100000
-	maxHours := 15000000
-	coins := make([]uint64, maxCoinsLen)
-	for i := 0; i < iterations; i++ {
-		coinsLen := rand.Intn(maxCoinsLen) + 1
-
-		maxCoins := rand.Intn(maxMaxCoins) + 1
-
-		var totalCoins uint64
-		for i := 0; i < coinsLen; i++ {
-			coins[i] = uint64(rand.Intn(maxCoins) + 1)
-
-			var err error
-			totalCoins, err = mathutil.AddUint64(totalCoins, coins[i])
-			require.NoError(t, err)
-		}
-
-		hours := uint64(rand.Intn(maxHours))
-
-		output, err := DistributeCoinHoursProportional(coins[:coinsLen], hours)
-		require.NoError(t, err)
-
-		require.Equal(t, coinsLen, len(output))
-
-		var totalHours uint64
-		for _, h := range output {
-			if hours >= totalCoins {
-				require.NotEqual(t, uint64(0), h)
-			}
-
-			var err error
-			totalHours, err = mathutil.AddUint64(totalHours, h)
-			require.NoError(t, err)
-		}
-
-		require.Equal(t, hours, totalHours)
-	}
-}
-
 func TestWalletValidate(t *testing.T) {
 	goodMetaUnencrypted := map[string]string{
 		"filename":  "foo.wlt",
@@ -2705,347 +1268,5 @@ func TestWalletValidate(t *testing.T) {
 			err := w.Validate()
 			require.Equal(t, tc.err, err)
 		})
-	}
-}
-
-func TestSignTransaction(t *testing.T) {
-	txnSigned, uxs, seckeys := makeTransaction(t, 4)
-	require.Equal(t, 4, len(uxs))
-	require.Equal(t, 4, len(seckeys))
-	txnUnsigned := txnSigned
-	txnUnsigned.Sigs = make([]cipher.Sig, len(txnSigned.Sigs))
-
-	w := &Wallet{}
-	for _, x := range seckeys {
-		p := cipher.MustPubKeyFromSecKey(x)
-		a := cipher.AddressFromPubKey(p)
-		err := w.AddEntry(Entry{
-			Address: a,
-			Public:  p,
-			Secret:  x,
-		})
-		require.NoError(t, err)
-
-		// Add unrelated entries
-		err = w.AddEntry(makeEntry())
-		require.NoError(t, err)
-	}
-
-	badTxnInnerHash, _, _ := makeTransaction(t, 2)
-	badTxnInnerHash.InnerHash = testutil.RandSHA256(t)
-
-	badTxnNoInputs, _, _ := makeTransaction(t, 2)
-	badTxnNoInputs.Sigs = make([]cipher.Sig, len(badTxnNoInputs.Sigs))
-	badTxnNoInputs.In = nil
-	err := badTxnNoInputs.UpdateHeader()
-	require.NoError(t, err)
-
-	badTxnNoSigs, _, _ := makeTransaction(t, 2)
-	badTxnNoSigs.Sigs = nil
-	err = badTxnNoSigs.UpdateHeader()
-	require.NoError(t, err)
-
-	txnOtherWallet, uxsOtherWallet, secKeysOtherWallet := makeTransaction(t, 4)
-
-	txnPartiallySigned := txnOtherWallet
-	err = txnPartiallySigned.Verify()
-	require.NoError(t, err)
-	err = txnPartiallySigned.VerifyInputSignatures(uxsOtherWallet)
-	require.NoError(t, err)
-
-	txnPartiallySigned.Sigs = make([]cipher.Sig, len(txnOtherWallet.Sigs))
-	copy(txnPartiallySigned.Sigs, txnOtherWallet.Sigs)
-	txnPartiallySigned.Sigs[1] = cipher.Sig{}
-	txnPartiallySigned.Sigs[2] = cipher.Sig{}
-	err = txnPartiallySigned.UpdateHeader()
-	require.NoError(t, err)
-
-	txnPartiallySigned2 := txnPartiallySigned
-	txnPartiallySigned2.Sigs = make([]cipher.Sig, len(txnPartiallySigned.Sigs))
-	copy(txnPartiallySigned2.Sigs, txnPartiallySigned.Sigs)
-	err = txnPartiallySigned2.UpdateHeader()
-	require.NoError(t, err)
-
-	txnOtherWallet.Sigs = make([]cipher.Sig, len(txnOtherWallet.Sigs))
-	err = txnOtherWallet.UpdateHeader()
-	require.NoError(t, err)
-
-	otherWallet := &Wallet{}
-	for i := 1; i < 3; i++ {
-		p := cipher.MustPubKeyFromSecKey(secKeysOtherWallet[i])
-		a := cipher.AddressFromPubKey(p)
-		err := otherWallet.AddEntry(Entry{
-			Address: a,
-			Public:  p,
-			Secret:  secKeysOtherWallet[i],
-		})
-		require.NoError(t, err)
-	}
-
-	cases := []struct {
-		name        string
-		w           *Wallet
-		txn         coin.Transaction
-		signIndexes []int
-		uxOuts      []coin.UxOut
-		err         error
-		partial     bool
-		complete    bool
-	}{
-		{
-			name:   "signed txn",
-			w:      w,
-			txn:    txnSigned,
-			uxOuts: uxs,
-			err:    NewError(errors.New("Transaction is fully signed")),
-		},
-
-		{
-			name: "bad txn inner hash",
-			w:    w,
-			txn:  badTxnInnerHash,
-			err:  NewError(errors.New("Transaction inner hash does not match computed inner hash")),
-		},
-
-		{
-			name: "txn no inputs",
-			w:    w,
-			txn:  badTxnNoInputs,
-			err:  NewError(errors.New("No transaction inputs to sign")),
-		},
-
-		{
-			name: "txn no sigs",
-			w:    w,
-			txn:  badTxnNoSigs,
-			err:  NewError(errors.New("Transaction signatures array is empty")),
-		},
-
-		{
-			name:   "len uxouts does not match len inputs",
-			w:      w,
-			txn:    txnUnsigned,
-			uxOuts: uxs[:2],
-			err:    errors.New("len(uxOuts) != len(txn.In)"),
-		},
-
-		{
-			name:        "too many sign indexes",
-			w:           w,
-			txn:         txnUnsigned,
-			uxOuts:      uxs,
-			signIndexes: []int{0, 1, 2, 3, 4, 5},
-			err:         NewError(errors.New("Number of signature indexes exceeds number of inputs")),
-		},
-
-		{
-			name:        "sign index out of range",
-			w:           w,
-			txn:         txnUnsigned,
-			uxOuts:      uxs,
-			signIndexes: []int{0, 1, 5, 2},
-			err:         NewError(errors.New("Signature index out of range")),
-		},
-
-		{
-			name:        "duplicate value in sign indexes",
-			w:           w,
-			txn:         txnUnsigned,
-			uxOuts:      uxs,
-			signIndexes: []int{0, 1, 1},
-			err:         NewError(errors.New("Duplicate value in signature indexes")),
-		},
-
-		{
-			name:   "wallet cannot sign any input",
-			w:      w,
-			txn:    txnOtherWallet,
-			uxOuts: uxsOtherWallet,
-			err:    NewError(errors.New("Wallet cannot sign all requested inputs")),
-		},
-
-		{
-			name:   "wallet cannot sign some inputs",
-			w:      otherWallet,
-			txn:    txnOtherWallet,
-			uxOuts: uxsOtherWallet,
-			err:    NewError(errors.New("Wallet cannot sign all requested inputs")),
-		},
-
-		{
-			name:        "wallet cannot sign all specified inputs",
-			w:           otherWallet,
-			txn:         txnOtherWallet,
-			uxOuts:      uxsOtherWallet,
-			signIndexes: []int{2, 0},
-			err:         NewError(errors.New("Wallet cannot sign all requested inputs")),
-		},
-
-		{
-			name:   "valid unsigned txn, all sigs",
-			w:      w,
-			txn:    txnUnsigned,
-			uxOuts: uxs,
-			err:    nil,
-		},
-
-		{
-			name:        "valid unsigned txn, some sigs defined",
-			w:           w,
-			txn:         txnUnsigned,
-			signIndexes: []int{1, 2},
-			uxOuts:      uxs,
-			err:         nil,
-		},
-
-		{
-			name:        "valid unsigned txn, all sigs defined",
-			w:           w,
-			txn:         txnUnsigned,
-			signIndexes: []int{0, 1, 2, 3},
-			uxOuts:      uxs,
-			err:         nil,
-		},
-
-		{
-			name:        "valid unsigned txn, all sigs defined, unordered",
-			w:           w,
-			txn:         txnUnsigned,
-			signIndexes: []int{2, 1, 3, 0},
-			uxOuts:      uxs,
-			err:         nil,
-		},
-
-		{
-			name:        "valid, wallet can sign the specified inputs, but not others",
-			w:           otherWallet,
-			txn:         txnOtherWallet,
-			uxOuts:      uxsOtherWallet,
-			signIndexes: []int{2},
-			err:         nil,
-		},
-
-		{
-			name:        "valid, transaction partially signed, unfinished signing",
-			w:           otherWallet,
-			txn:         txnPartiallySigned,
-			uxOuts:      uxsOtherWallet,
-			signIndexes: []int{2},
-			err:         nil,
-			partial:     true,
-			complete:    false,
-		},
-
-		{
-			name:        "valid, transaction partially signed, finished signing",
-			w:           otherWallet,
-			txn:         txnPartiallySigned2,
-			uxOuts:      uxsOtherWallet,
-			signIndexes: []int{1, 2},
-			err:         nil,
-			partial:     true,
-			complete:    true,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			signedTxn, err := tc.w.SignTransaction(&tc.txn, tc.signIndexes, tc.uxOuts)
-			if tc.err != nil {
-				require.Equal(t, tc.err, err)
-				return
-			}
-
-			require.NoError(t, err)
-
-			// The original txn should not be modified
-			require.False(t, reflect.DeepEqual(tc.txn, *signedTxn))
-
-			if len(tc.signIndexes) == 0 || len(tc.signIndexes) == len(tc.uxOuts) || tc.complete {
-				// Transaction should be fully signed
-				require.False(t, signedTxn.IsFullyUnsigned())
-				err = signedTxn.Verify()
-				require.NoError(t, err)
-				err = signedTxn.VerifyInputSignatures(tc.uxOuts)
-				require.NoError(t, err)
-			} else {
-				// index of a valid signature should be found in the signIndexes
-				for _, x := range tc.signIndexes {
-					require.False(t, signedTxn.Sigs[x].Null())
-				}
-
-				if !tc.partial {
-					// Number of signatures should equal length of signIndexes
-					nSigned := 0
-					for _, s := range signedTxn.Sigs {
-						if !s.Null() {
-							nSigned++
-						}
-					}
-
-					require.Equal(t, len(tc.signIndexes), nSigned)
-				}
-			}
-		})
-	}
-}
-
-func makeTransaction(t *testing.T, nInputs int) (coin.Transaction, []coin.UxOut, []cipher.SecKey) {
-	txn := coin.Transaction{}
-
-	toSign := make([]cipher.SecKey, 0)
-	uxs := make([]coin.UxOut, 0)
-	for i := 0; i < nInputs; i++ {
-		ux, s := makeUxOutWithSecret(t)
-		err := txn.PushInput(ux.Hash())
-		require.NoError(t, err)
-		toSign = append(toSign, s)
-		uxs = append(uxs, ux)
-	}
-
-	err := txn.PushOutput(makeAddress(), 1e6, 50)
-	require.NoError(t, err)
-	err = txn.PushOutput(makeAddress(), 5e6, 50)
-	require.NoError(t, err)
-	txn.SignInputs(toSign)
-	err = txn.UpdateHeader()
-	require.NoError(t, err)
-
-	return txn, uxs, toSign
-}
-
-func makeUxOutWithSecret(t *testing.T) (coin.UxOut, cipher.SecKey) {
-	body, sec := makeUxBodyWithSecret(t)
-	return coin.UxOut{
-		Head: coin.UxHead{
-			Time:  100,
-			BkSeq: 2,
-		},
-		Body: body,
-	}, sec
-}
-
-func makeUxBodyWithSecret(t *testing.T) (coin.UxBody, cipher.SecKey) {
-	p, s := cipher.GenerateKeyPair()
-	return coin.UxBody{
-		SrcTransaction: testutil.RandSHA256(t),
-		Address:        cipher.AddressFromPubKey(p),
-		Coins:          1e6,
-		Hours:          100,
-	}, s
-}
-
-func makeAddress() cipher.Address {
-	p, _ := cipher.GenerateKeyPair()
-	return cipher.AddressFromPubKey(p)
-}
-
-func makeEntry() Entry {
-	p, s := cipher.GenerateKeyPair()
-	a := cipher.AddressFromPubKey(p)
-	return Entry{
-		Secret:  s,
-		Public:  p,
-		Address: a,
 	}
 }
